Unit Test impl. Line and Branch weighted average  Coverage %85 over

+++ policyService/policyService/src/test/java/PolicyProject/policyService/interfaces/mappers/PolicyMapperTest/EarthQuakeMapperTest.java
@@ -43,7 +43,7 @@ public class EarthQuakeMapperTest {
         assertEquals(request.policyStartDate(), model.policyStartDate());
         assertEquals(request.policyEndDate(), model.policyEndDate());
         assertEquals(request.tckn(), model.tckn());
-        assertEquals(request.policyAmount(), model.policyAmount());
+        assertEquals(request.policyAmount().doubleValue(), model.policyAmount().doubleValue());
     }
     @Test
     public void testCreateEarthQuakeRequestToEarthQuakeModel_withNullValues() {
